pyproject.toml
[tool.poetry]
name = "microshop"
version = "0.1.0"
description = ""
authors = ["Азат <yanberdin.az@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
